feat: refactor sub components
